airflow/pyproject.toml
[project]
name = "airflow-data-catalog"
version = "1.0"
description = "Workflow orchestration with Airflow"
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
    "apache-airflow>=2.10.5,<2.11",                     # pin to 2.x to prevent airbyte-api from upgrading
    "sqlalchemy>=1.4.54,<2.0",                          # Airflow doesn't yet support SQLAlchemy 2.x
    "psycopg2-binary==2.9.11",                          # psycopg3 is only supported by SQLAlchemy 2.0+

    "apache-airflow-providers-airbyte==5.2.5",          # Airbyte operator
    "apache-airflow-providers-docker>=4.4.5,<4.5",      # Docker operator for running containers
    "acryl-datahub-airflow-plugin[airflow2]==1.3.0",    # DataHub extension for OpenLineage (match Server version)
]

[project.optional-dependencies]
celery = [
    "apache-airflow-providers-celery>=3.10.0,<4.0",    # required for CeleryExecutor
    "apache-airflow-providers-redis>=3.8.0,<4.0",      # required for CeleryExecutor
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
    "jupyterlab>=4.5.3",
]

[tool.setuptools.packages.find]
exclude = ["config", "dags", "logs", "plugins"]

[tool.ruff]
line-length = 100

[tool.ruff.format]
skip-magic-trailing-comma = false
preview = true  # unstable: enables 'hug_parens_with_braces_and_square_brackets' for compact StructType([...])

[tool.ruff.lint]
preview = true  # unstable: enables latest lint rules before stabilization
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort (import sorting)
    "UP",   # pyupgrade (modernize syntax for newer Python)
    "B",    # flake8-bugbear (common bug patterns)
    "SIM",  # flake8-simplify (code simplification)
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",    # line too long (handled by formatter)
    "RUF013",  # implicit Optional
]
fixable = ["ALL"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402"]  # allow imports not at top in notebooks
